Fast game scene (#26)
* update

* GameScene: perf, restore fixes, loading debounce, standard orbit camera (#25)

* perf: cache TooltipDetails by element identity in createMetadata

createMetadata() iterated 6500 blocks calling tooltipResolver.resolve()
per position. CandidateCacheKey included (x,y,z), causing 6499 unique
keys that all missed the 512-entry LRU cache. Each call did tier loops
and reflection tree walks.

Add IdentityHashMap<IStructureElement, TooltipDetails> cache keyed by
element instance identity at the createMetadata() loop level. Same
element instance shared across positions hits cache on first lookup.

Also strip SURVIVAL_CONSTRUCT_OPTION from initial buildPreviewSelection
in doScan() to avoid pointless survival→creative fallback on fake
GuidebookLevel.

Before: doScan bindingMs=3028, createMetadata.resolveTotalMs=2887
After:  doScan bindingMs=121,  createMetadata.resolveTotalMs=5

* fix: channel slider disappears and stops responding after cache restore

Three root causes in the StructureLib scene cache restore path:

1. Slider disappears on page revisit:
   captureInitialStructureStateIfAbsent() ran before ANALYZE_LIMITS
   completed, caching 0-channel metadata. Force re-capture after
   ANALYZE_LIMITS enriches metadata with channel data.

2. Slider drag has no effect after cache restore:
   resolvePreviewBindingKey() uses object identity (==) to match
   binding to runtime state. After restoreInto() recreates bindings,
   the old identity check fails. Add rebindPreviewRuntimeStates()
   to update state.binding references after restore.
   Also restore hasRebuildRecipe from baseRequest.

3. Slider drag causes lag spikes:
   captureInitialStructureStateIfAbsent() was forced on every
   BUILD_SELECTION result (248ms for 6500 blocks). Only force
   re-capture on ANALYZE_LIMITS; BUILD results use the original
   null-guarded capture path.

Add build deduplication: skip queueStructureLibSelectionBuild when
buildPending is already true. Resubmit after build completes if
pending selection changed during flight.

* fix: Loading indicator flicker — debounce was bypassed by render code

isLoading() method had a 200ms debounce but drawLoadProgressOverlay()
read the raw isLoading field directly, bypassing the debounce entirely.
Also ensure setLoadFailure() calls setLoading() so the debounce timestamp
is properly reset.

* refactor: standard orbit camera for pan and rotate

Replace offset-based pan with standard orbit camera model:
- View matrix: V = T(offset) · R · T(-rc)  (R = Rz·Rx·Ry)
- Projection: simple zoom without offset-centering
- Pan: move rotationCenter directly via R⁻¹·(dx/s, -dy/s, 0)
- Rotate: orbits around rotationCenter automatically
- endDrag: no consolidation needed (pan/rotate are orthogonal)

Removed the post-rotation T(+rc) from the view matrix which was
the root cause of (I-R) attenuation making pan formulas complex.

* fix ae2 cable, remove bootstrap

* idea clean

* update multiBlockBase.checkMachine

* remove StructureLibPreviewBootstrap

* undo math clamp

* Cache discoverNativeNamespaceRoots results to avoid repeated listFiles() on Windows

* refactor: unify StructureLib cache — definition-based, no LRU, no worker

Replace 3 LRU caches (IMPORT_RESULT_CACHE, CONTROL_ANALYSIS_CACHE,
ANALYSIS_SNAPSHOT_CACHE) and async StructureLibPreviewWorker with:
- Mixin accessors for BlockRenderer6343 ConstructableData + multiblock registry
- StructureLibDefinitionCache: full Map<IConstructable, ConstructableData>, no eviction
- Synchronous facade.importScene() driven by trigger ItemStack
- Per-binding footprint tracking in LytGuidebookScene for diff-update

Deleted: StructureLibPreviewWorker/Task/Result, ImportCacheKey, BoundedCache
Eliminated: initialStructureState, structureLibBaseState snapshots

* wip: StructureLib integration simplification checkpoint

- Simplify import paths by removing unused direct block-level operations from
  SceneScript and LytGuidebookScene (delegated to StructureLibRuntimeFacade)
- Unify StructureLibDefinitionCache to single definition-based lookup
- Clean up deprecated delegate methods in StructureLibRuntimeFacade

This is a WIP checkpoint before rewriting the StructureLib integration layer
to address the dual-state synchronization issue with Slider rebuild paths.

* wip: rewrite StructureLib integration layer

- Replace StructureLibRuntimeFacade (1250 lines, 9 inner classes) with
  StructureLibBuildService (~280 lines, single error boundary)
- Merge 4 overlapping DTOs into StructureLibBuildRequest + BuildResult
- Delete StructureLibElementTooltipResolver, PreviewMetadataFactory,
  VisitCollector, PreviewItemSource (data already in build result)
- Simplify StructureLibSceneMetadata (421 -> 200 lines, remove tooltip fill)
- Simplify StructureLibSceneBinding (remove dual-state sync patterns)
- Rewrite ImportStructureLibElementCompiler as pure parse (no world ops)
- Update rebuildStructureLib to use BuildService + sceneLevel.clear()
- Update all callers (SceneScript, SemanticProvider, export tools, editor)
- Add INEIPreviewModifier hooks (matching BlockRenderer6343 behavior)
- Fix: PLACE_HATCHES_OPTION restores gt_hatch channel wiring
- Fix: Channel slider defaults to 0 (NOT_SET) instead of max value
- Fix: survivalConstruct uses same trigger per round (no copy)
- Fix: IItemSource fallback scans full item registry

Net: ~3300 lines removed

* fix: snapshotBlocks min-corner, use CreativeItemSource, remove survival option

* wip: symmetric build/clear/rebuild for ImportStructure + StructureLib

- Add SnbtPlacement to record ImportStructure placement config
- Extract placeStructure() static method from ImportStructureElementCompiler
- Change compile() to register-only (place via scene.build())
- Add build(), rebuild(), clear() to LytGuidebookScene
- Rewrite SceneScript state machine: INIT -> AWAIT_SNBT -> BUILD
- Remove duplicate registerDefaultStructureBinding code
- Deprecate rebuildStructureLib() in favor of rebuild()

* fix: remove resetSceneToInitialState — scene is self-contained, re-entry should not rebuild blocks

resetSceneToInitialState() was calling clear() + build() on every page
re-entry via prepareForDisplay(). The scene is already a self-contained
node in the LytDocument tree with blocks persisted in its level. Re-entry
only needs to restore interactive state (drag, hover, slider positions),
not clear and rebuild the entire level from placement configs.

* feat: keep user interactive state across page re-entry

* idea clean

* remove ConcurrentSceneCacheManager and clean

* remove GameSceneConcurrentManager and prewarm — dead code after cache layer removal

* spotlessApply

---------

+++ src/main/java/com/hfstudio/guidenh/guide/document/block/LytContentTabsBlock.java
@@ -349,11 +349,11 @@ public class LytContentTabsBlock extends LytBlock implements InteractiveElement,
         }
 
         // Each tab button
-        int tabWidth = (int) (headerBounds.width() / tabs.size());
+        int tabWidth = headerBounds.width() / tabs.size();
         for (int i = 0; i < tabs.size(); i++) {
             TabState tab = tabs.get(i);
-            int tabX = (int) headerBounds.x() + (i * tabWidth);
-            LytRect tabBounds = new LytRect(tabX, (int) headerBounds.y(), tabWidth, (int) headerBounds.height());
+            int tabX = headerBounds.x() + (i * tabWidth);
+            LytRect tabBounds = new LytRect(tabX, headerBounds.y(), tabWidth, headerBounds.height());
 
             String extra = "Index: " + i;
             if (i == selectedIndex) {
